% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lsq_plot}
\alias{lsq_plot}
\title{Plot cumulative distribution
\code{lsq_plot} Plots the cumulative distribution of the data as well as the best fit linear model line.}
\usage{
lsq_plot(object)
}
\arguments{
\item{object}{neutrality test object}
}
\value{
ggplot object.
}
\description{
Plot cumulative distribution
\code{lsq_plot} Plots the cumulative distribution of the data as well as the best fit linear model line.
}
\examples{
lsq_plot(neutralitytest(VAFselection, fmin = 0.1, fmax = 0.25))
}
