% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutralitytestr.R
\docType{package}
\name{neutralitytestr}
\alias{neutralitytestr}
\alias{neutralitytestr-package}
\title{\code{neutralitytestr} package}
\description{
Package to test a neutral evolutionary model on deep sequencing data.
}
\details{
See the README on
\href{https://github.com/marcjwilliams1/neutralitytestr#readme}{GitHub}
}
