% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{normalized_plot}
\alias{normalized_plot}
\title{Plot normalized cumulative distribution
\code{normalized_plot} Plots the (normalized) cumulative distribution of the data as well as the theoretical expectation from
a neutral evolutionary model.}
\usage{
normalized_plot(object)
}
\arguments{
\item{object}{neutrality test object}
}
\value{
ggplot object.
}
\description{
Plot normalized cumulative distribution
\code{normalized_plot} Plots the (normalized) cumulative distribution of the data as well as the theoretical expectation from
a neutral evolutionary model.
}
\examples{
normalized_plot(neutralitytest(VAFselection, fmin = 0.1, fmax = 0.25))
}
