% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_all}
\alias{plot_all}
\title{Plot all plots in the package and make composite figure.
\code{plot_all} Plots histogram, lqs plot and normalized plot and plot all togetger.}
\usage{
plot_all(object)
}
\arguments{
\item{object}{neutrality test object}
}
\value{
ggplot object.
}
\description{
Plot all plots in the package and make composite figure.
\code{plot_all} Plots histogram, lqs plot and normalized plot and plot all togetger.
}
\examples{
plot_all(neutralitytest(VAFselection, fmin = 0.1, fmax = 0.25))
}
