% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{vaf_histogram}
\alias{vaf_histogram}
\title{Plot VAF histogram
\code{vaf_histogram} Plots a histogram of the variant allele frequencies.}
\usage{
vaf_histogram(object)
}
\arguments{
\item{object}{neutrality test object}
}
\value{
ggplot object.
}
\description{
Plot VAF histogram
\code{vaf_histogram} Plots a histogram of the variant allele frequencies.
}
\examples{
vaf_histogram(neutralitytest(VAFselection, fmin = 0.1, fmax = 0.25))
}
