% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VAFneutral}
\alias{VAFneutral}
\title{Synthetic sequencing data generated from a evolutionary based cancer simulation.}
\format{
A vector with variant allele frequencies (VAFs) ranging from 0 to 1
}
\source{
Generated using cancer sequencing simulation \url{https://github.com/marcjwilliams1/CancerSeqSim.jl}
}
\usage{
VAFneutral
}
\description{
This data is generated from a neutral evolutionary model where all subclonal mutations are neutral.
}
\keyword{datasets}
