% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VAFselection}
\alias{VAFselection}
\title{Synthetic sequencing data generated from a evolutionary based cancer simulation.}
\format{
A vector with variant allele frequencies (VAFs) ranging from 0 to 1
}
\source{
Generated using cancer sequencing simulation \url{https://github.com/marcjwilliams1/CancerSeqSim.jl}
}
\usage{
VAFselection
}
\description{
This data is generated from an evolutionary model where there is on subclonal population
and all other mutations are neutral passengers.
}
\keyword{datasets}
